Add ./configure --enable-fontforge-extras

+++ contrib/cidmap/src/mapjapan2.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include "utype.h"
 
 static char used[0x110000];
 static int cid_2_unicode[0x10000];
@@ -86,7 +87,7 @@ return( -1 );
 		mults[j++] = vals[i];
 	}
 	if ( j>MULT_MAX ) {
-	    fprintf( stderr, "Too many multiple values for %04x, need %d slots\n", val, j );
+	    fprintf( stderr, "Too many multiple values for %04x, need %d slots\n", (unsigned int)(val), j );
 exit(1);
 	}
     }
@@ -118,7 +119,7 @@ int main(int argc, char **argv) {
 	    used[uni] = 1;
 	    cid_2_unicode[cid] = uni;
 	} else {
-	    sprintf( buffer, "uni%04X.dup%d", uni, ++used[uni] );
+	    sprintf( buffer, "uni%04X.dup%d", (unsigned int)(uni), ++used[uni] );
 	    nonuni_names[cid] = strdup(buffer);
 	}
 	max = cid;
@@ -131,14 +132,14 @@ int main(int argc, char **argv) {
 	    for ( i=1; cid+i<=max && cid_2_unicode[cid+i]==cid_2_unicode[cid]+i && cid_2_unicodemult[cid+i][0]==0; ++i );
 	    --i;
 	    if ( i!=0 ) {
-		printf( "%d..%d %04x\n", cid, cid+i, cid_2_unicode[cid] );
+		printf( "%d..%d %04x\n", cid, cid+i, (unsigned int)(cid_2_unicode[cid]) );
 		cid += i;
 	    } else
-		printf( "%d %04x\n", cid, cid_2_unicode[cid] );
+		printf( "%d %04x\n", cid, (unsigned int)(cid_2_unicode[cid]) );
 	} else if ( cid_2_unicode[cid]!=-1 ) {
-	    printf( "%d %04x", cid, cid_2_unicode[cid]);
+	    printf( "%d %04x", cid, (unsigned int)(cid_2_unicode[cid]) );
 	    for ( i=0; cid_2_unicodemult[cid][i]!=0; ++i )
-		printf( ",%04x", cid_2_unicodemult[cid][i]);
+		printf( ",%04x", (unsigned int)(cid_2_unicodemult[cid][i]) );
 	    printf( "\n");
 	} else if ( nonuni_names[cid]!=NULL )
 	    printf( "%d /%s\n", cid, nonuni_names[cid] );
